add a flag to retrieve (or not) entities recursively
When retrieving linked entities, we can follow multiple hops of links to collect all data related to an entity, however a) it may not be required often, b) when links are cyclic between entities of the same type (e.g. person to person), it will end up in an infinite recursion. The issues does not occur with entities of different types, as links are not bi-directional and with the current design only certain types of links can be established.

Further fix to mis-ordered fields in entity to location crossref table