optimise deployment configs
Owner password is now expected already bcrypt-hashed instead of being stored in plaintext. Currently the same hash of the password is used for database user password as it needs to be stored in database config in plaintext.

Minimum and maximum heap size Java config variables are used to reduce memory memory footprint of an individual HAT (at launch with default config it uses ~20MB of heap out of allocted 300MB).