enable HAT to migrate its own database schema
for HATs running live making database changes is very difficult due to the number of HATs in existence as well as due to the fact that an organisation running the HATs is not supposed to have access to the database and hence can not modify it.

Introducing liquibase for HAT to detect own database evolutions and apply them at boot time. Using the mechanism to add a notion of applications (HAT Services) to the database.